+++ src/components/index.ts
@@ -3,6 +3,7 @@ export * from "./card";
 export * from "./download-item";
 export * from "./episode-card";
 export * from "./expandable-section";
+export * from "./form";
 export * from "./loading";
 export * from "./poster-card";
 export * from "./problem-set-card";
