+++ src/routes/courses/index.tsx
@@ -1,9 +1,3 @@
-import { FormComponent } from "../../components/reusable-form/form.stories";
-
 export const Courses = () => {
-  return (
-    <div className="flex justify-center">
-      <FormComponent />
-    </div>
-  );
+  return <div>This is the courses page</div>;
 };
